Cleaned directory structure, and modified jobs.
I made a directory for source (src) and one for includes (include).  This will k
eep the project tidy as it expands.

I also modified the way jobs work so now you fill out a struct (job_t) and pass
that to the add job function.  This allows more flexibility in the way jobs work
.