Wrote an I2C driver.
This driver is non-blocking and interrupt driven.  There are a few things left to do such as adding multiple I2C module support (this is partly completed, and is awaiting a FIFO interface to make this easier).

Using the included examples it is possible to easier read from the EEPROM or write to it.  Unlike the microchip example, this will write and read more than one byte!