Modified the I2C driver.
The frames are now created by the user of the driver, and contain the status of that frame and also the transmit and receive buffers.  This means that the user of the driver can access their data even if another transaction starts (which previously would have resulted in the data being overwritten).